[app] Translate about page
