+++ components/molecules/details-card.tsx
@@ -1,12 +1,12 @@
 import Icon from '../atoms/icon'
 
-type Card = {
+export type DetailsCardProps = {
   icon: string
   title: string
   items: Array<string>
 }
 
-export default function DetailsCard ({ icon, title, items }: Card) {
+export default function DetailsCard ({ icon, title, items }: DetailsCardProps) {
   return (
     <div className='h-full bg-white pt-16 pr-[15%] pb-24 pl-12 z-10'>
       <Icon src={icon} alt={title} width={0} height={0}
