+++ components/organisms/base-cards-wrapper.tsx
@@ -1,9 +1,8 @@
+import BaseCard, { BaseCardProps } from '../molecules/base-card'
 import Surtitle from '../atoms/surtitle'
-import { Card } from '../../data/base-cards'
-import BaseCard from '../molecules/base-card'
 
 type BaseCardsWrapperProps = {
-  cards: Array<Card>
+  cards: Array<BaseCardProps>
   text: string
   className?: string
 }
