+++ data/base-cards.ts
@@ -1,11 +1,4 @@
-export type Card = {
-  number: string
-  icon: string
-  title: string
-  className: string
-}
-
-export const baseCardsPartners: Array<Card> = [
+export const baseCardsPartners = [
   {
     number: '01',
     icon: '/images/cloud-icon.png',
@@ -29,32 +22,5 @@ export const baseCardsPartners: Array<Card> = [
     icon: '/images/devops-icon.png',
     title: 'Hashicorp',
     className: 'bg-orange-lighter'
-  }
-]
-
-export const baseCardsAbout: Array<Card> = [
-  {
-    number: '01',
-    icon: '/images/automation-icon.png',
-    title: 'Automation both Infrastructure & Applications',
-    className: 'bg-orange-medium'
-  },
-  {
-    number: '02',
-    icon: '/images/devops-icon.png',
-    title: 'DevOps Services',
-    className: 'bg-orange-lighter'
-  },
-  {
-    number: '03',
-    icon: '/images/cloud-icon.png',
-    title: 'Cloud',
-    className: 'bg-blue-light'
-  },
-  {
-    number: '04',
-    icon: '/images/monitoring-icon.png',
-    title: 'Monitoring & Security',
-    className: 'bg-orange-medium'
   }
 ]
