--- data/services.ts
@@ -1,21 +1,0 @@
-export type Detail = {
-  description: string
-}
-
-export const details: Array<Detail> = [
-  {
-    description: 'Learning about the Cloud and it\'s best practices'
-  },
-  {
-    description: 'Accumulated cloud knowledge in ready to use packages'
-  },
-  {
-    description: 'Customizable Infrastructure as Code (IaC) components'
-  },
-  {
-    description: 'Access to experts to guide the team on their journey'
-  },
-  {
-    description: 'Learning backwards approach to enable the team'
-  }
-]
