+++ pages/about.tsx
@@ -1,17 +1,19 @@
 import useTranslation from 'next-translate/useTranslation'
+import { BaseCardProps } from '../components/molecules/base-card'
+import { Card } from '../components/molecules/team-card'
 import Head from 'next/head'
 import Layout from '../components/organisms/layout'
 import MediumTitleIntro from '../components/molecules/medium-title-intro'
 import Gallery from '../components/molecules/gallery'
 import BaseCardsWrapper from '../components/organisms/base-cards-wrapper'
-import { baseCardsAbout } from '../data/base-cards'
 import BackgroundImage from '../components/atoms/background-image'
 import TeamCardsWrapper from '../components/organisms/team-cards-wrapper'
-import { teamCards } from '../data/team-cards'
 import Footer from '../components/molecules/footer'
 
 export default function AboutPage () {
   const { t } = useTranslation('about')
+  const cardsBase: Array<BaseCardProps> = t('about:focus', { count: 4 }, { returnObjects: true })
+  const cardsTeam: Array<Card> = t('about:team', { count: 3 }, { returnObjects: true })
 
   return (
     <>
@@ -23,10 +25,10 @@ export default function AboutPage () {
           textColLeft={t('introTextColLeft')}
           textColRight={t('introTextColRight')} />
         <Gallery />
-        <BaseCardsWrapper cards={baseCardsAbout} text={t('cardsWrapperSurtitle')} className='after:w-64' />
+        <BaseCardsWrapper cards={cardsBase} text={t('cardsWrapperSurtitle')} className='after:w-64' />
         <BackgroundImage src='/images/divider-dark.png' alt='Plants divider' width={1920} height={0}
           className='-mt-[7rem] sm:-mt-36 md:-mt-44 lg:-mt-56 xl:-mt-72 2xl:-mt-96' />
-        <TeamCardsWrapper cards={teamCards} />
+        <TeamCardsWrapper cards={cardsTeam} />
       </Layout>
       <footer>
         <Footer version='blue' />
