+++ pages/training.tsx
@@ -1,4 +1,5 @@
 import useTranslation from 'next-translate/useTranslation'
+import { DetailsCardProps } from '../components/molecules/details-card'
 import Head from 'next/head'
 import Layout from '../components/organisms/layout'
 import BackgroundImage from '../components/atoms/background-image'
@@ -10,8 +11,7 @@ import Footer from '../components/molecules/footer'
 
 export default function TrainingPage () {
   const { t } = useTranslation('training')
-  const cards: Array<{icon: string; title: string; items: Array<string>}> =
-    t('training:features', { count: 4 }, { returnObjects: true })
+  const cardsDetails: Array<DetailsCardProps> = t('training:features', { count: 4 }, { returnObjects: true })
   const details: Array<string> = t('training:details', { count: 5 }, { returnObjects: true })
 
   return (
@@ -30,7 +30,7 @@ export default function TrainingPage () {
       <BackgroundImage src='/images/mountain-background.png' alt={t('altMainIllustration')}
         className='-mt-[10rem] sm:-mt-[15rem] md:-mt-[17rem] lg:-mt-[30rem] xl:-mt-[40rem] 2xl:-mt-[50rem]'
         width={1920} height={0} />
-      <DetailsCardsWrapper cards={cards} />
+      <DetailsCardsWrapper cards={cardsDetails} />
       <BackgroundImage src='/images/divider-light.png' alt={t('altDivider')} className='-mt-40 xl:-mt-64'
         width={1920} height={0} />
       <Contact />
